2020_0316 Modified for solving the +0.5 pixel
Because of the +0.5 pixel shift problem, the program was renewaled.
Ploblem of projection onto sinusoidal or eqr was also solved in this time.
Old program become old_xxxxx.py and new one is sgli_conv_xxx.py.
After June, 2021, we should use old one as +0.5 shift problem will be
solved.